+++ test/Raven.CodeAnalysis.Tests/CodeGen/UserDefinedOperatorCodeGenTests.cs
@@ -1,0 +1,62 @@
+using System;
+using System.IO;
+using System.Reflection;
+
+using Raven.CodeAnalysis.Syntax;
+
+namespace Raven.CodeAnalysis.Tests;
+
+public class UserDefinedOperatorCodeGenTests
+{
+    [Fact]
+    public void BinaryExpression_InvokesUserDefinedEqualityOperator()
+    {
+        var code = """
+class Number {
+    val Value: int
+
+    init(value: int) {
+        Value = value
+    }
+
+    static func ==(left: Number, right: Number) -> bool {
+        return left.Value == right.Value
+    }
+
+    static func !=(left: Number, right: Number) -> bool {
+        return !(left == right)
+    }
+}
+
+class Runner {
+    public func Equal() -> bool {
+        return Number(1) == Number(1)
+    }
+
+    public func NotEqual() -> bool {
+        return Number(1) != Number(1)
+    }
+}
+""";
+
+        var syntaxTree = SyntaxTree.ParseText(code);
+        var references = TestMetadataReferences.Default;
+
+        var compilation = Compilation.Create("test", new CompilationOptions(OutputKind.DynamicallyLinkedLibrary))
+            .AddSyntaxTrees(syntaxTree)
+            .AddReferences(references);
+
+        using var peStream = new MemoryStream();
+        var result = compilation.Emit(peStream);
+        Assert.True(result.Success, string.Join(Environment.NewLine, result.Diagnostics));
+
+        using var loaded = TestAssemblyLoader.LoadFromStream(peStream, references);
+        var type = loaded.Assembly.GetType("Runner", throwOnError: true)!;
+        var instance = Activator.CreateInstance(type)!;
+        var equal = type.GetMethod("Equal", BindingFlags.Instance | BindingFlags.Public)!;
+        var notEqual = type.GetMethod("NotEqual", BindingFlags.Instance | BindingFlags.Public)!;
+
+        Assert.True((bool)equal.Invoke(instance, Array.Empty<object>())!);
+        Assert.False((bool)notEqual.Invoke(instance, Array.Empty<object>())!);
+    }
+}